+++ MvvmPlugin/MvvmPlugin/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.7.0
+- Added Convert to Relay Command context action to convert to source generated RelayCommand
+- Create Relay Context action to generate relay command when Source generation is not available
+
 ## 0.6.3
 - Fixed issue with using Other and identifying candidates that are not direct subtypes of ObservableObject